moving around shared code
- moved code that is shared between the gui and the game into a shared
package so that it can be easily reused by other projects
