+++ src/main/java/net/demilich/metastone/gui/playmode/GameContextVisualizable.java
@@ -3,7 +3,7 @@ package net.demilich.metastone.gui.playmode;
 import java.util.ArrayList;
 import java.util.List;
 
-import net.demilich.metastone.AppConfig;
+import net.demilich.metastone.shared.AppConfig;
 import net.demilich.metastone.ApplicationFacade;
 import net.demilich.metastone.GameNotification;
 import net.demilich.metastone.game.GameContext;
