+++ src/main/java/net/demilich/metastone/gui/playmode/config/PlayModeConfigMediator.java
@@ -8,7 +8,7 @@ import net.demilich.nittygrittymvc.interfaces.INotification;
 import net.demilich.metastone.GameNotification;
 import net.demilich.metastone.game.decks.Deck;
 import net.demilich.metastone.game.decks.DeckFormat;
-import net.demilich.metastone.gui.gameconfig.GameConfig;
+import net.demilich.metastone.shared.gameconfig.GameConfig;
 import net.demilich.metastone.gui.playmode.PlayModeMediator;
 
 public class PlayModeConfigMediator extends Mediator<GameNotification> {
