+++ src/main/java/net/demilich/metastone/gui/simulationmode/SimulateGamesCommand.java
@@ -20,8 +20,8 @@ import net.demilich.metastone.game.GameContext;
 import net.demilich.metastone.game.Player;
 import net.demilich.metastone.game.decks.DeckFormat;
 import net.demilich.metastone.game.logic.GameLogic;
-import net.demilich.metastone.gui.gameconfig.GameConfig;
-import net.demilich.metastone.gui.gameconfig.PlayerConfig;
+import net.demilich.metastone.shared.gameconfig.GameConfig;
+import net.demilich.metastone.shared.gameconfig.PlayerConfig;
 import net.demilich.metastone.utils.Tuple;
 
 public class SimulateGamesCommand extends SimpleCommand<GameNotification> {
