+++ src/main/java/net/demilich/metastone/gui/trainingmode/PerformTrainingCommand.java
@@ -16,7 +16,7 @@ import net.demilich.metastone.game.cards.CardSet;
 import net.demilich.metastone.game.decks.DeckFormat;
 import net.demilich.metastone.game.logic.GameLogic;
 import net.demilich.metastone.game.statistics.Statistic;
-import net.demilich.metastone.gui.gameconfig.PlayerConfig;
+import net.demilich.metastone.shared.gameconfig.PlayerConfig;
 
 public class PerformTrainingCommand extends SimpleCommand<GameNotification> {
