+++ src/main/java/net/demilich/metastone/shared/AppConfig.java
@@ -1,4 +1,4 @@
-package net.demilich.metastone;
+package net.demilich.metastone.shared;
 
 public final class AppConfig {
