+++ src/main/java/net/demilich/metastone/shared/gameconfig/PlayerConfig.java
@@ -1,4 +1,4 @@
-package net.demilich.metastone.gui.gameconfig;
+package net.demilich.metastone.shared.gameconfig;
 
 import net.demilich.metastone.game.behaviour.IBehaviour;
 import net.demilich.metastone.game.cards.HeroCard;
