+++ src/test/java/net/demilich/metastone/tests/MassTest.java
@@ -17,7 +17,7 @@ import net.demilich.metastone.game.decks.DeckFactory;
 import net.demilich.metastone.game.decks.DeckFormat;
 import net.demilich.metastone.game.entities.heroes.HeroClass;
 import net.demilich.metastone.game.logic.GameLogic;
-import net.demilich.metastone.gui.gameconfig.PlayerConfig;
+import net.demilich.metastone.shared.gameconfig.PlayerConfig;
 
 public class MassTest extends TestBase {
